api/requirements.txt
bcrypt==4.0.1
click==8.0.0
distlib==0.3.7
filelock==3.12.4
Flask==2.2.1
Flask-Cors==4.0.0
Flask-DotEnv==0.1.2
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.5.1
itsdangerous==2.0.1
Jinja2==3.0.0
MarkupSafe==2.1.1
marshmallow==3.0.0
marshmallow-sqlalchemy==0.21.0
mysql-connector-python==8.0.30
platformdirs==2.6.2
protobuf==3.20.1
python-dotenv==1.0.0
six==1.14.0
SQLAlchemy==1.3.13
termcolor==1.1.0
virtualenv==20.16.5
Werkzeug==2.2.0
